% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abc_hs_dropout}
\alias{abc_hs_dropout}
\title{Chicago high school dropout rates.}
\format{A data frame.
\describe{
  \item{Name_of_School}{Name of school}
  ...
}}
\usage{
abc_hs_dropout
}
\description{
Chicago high school droput rates from 1995 and associated variables
like average students per teach and percent low income.
}
\details{
The data is based on:

Morton, Felicia B. (1995). Charting a School¹s Course. Chicago.
    February, pp. 86-95.

Rodkin, Dennis. (1995). 10 Keys for Creating Top High Schools.
    Chicago. February, pp. 78-85.

There are 18 data columns.  The data.frame is
larger because it has extra info.

This is the data set used in simulations by the ABC (Adaptive Behavior
and Cognition) research group.
}
\keyword{datasets}

