% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{aggregatePredictPair}
\alias{aggregatePredictPair}
\title{Predicts with heuristics and criterion, returning all row pairs.}
\usage{
aggregatePredictPair(fitted_heuristic_list, test_data, goal_type, ...)
}
\arguments{
\item{fitted_heuristic_list}{A list of heuristics already fitted to data,
e.g. ttbModel.}

\item{test_data}{Data to try to predict; must match columns in fit.}

\item{goal_type}{String identifying the goal of the models and criterion.
'ProbGreater' indicates to use probGreater and predictRoot.
ChooseGreater' indicates to use correctGrater and predictPairInternal.}

\item{...}{Optionally additional row pair functions, e.g. rowIndexes().}
}
\value{
A one-row matrix of numbers from 0 to 1, meaning proportion
  correct.  Each column is named with the heuristic's class.
}
\description{
Predicts with heuristics and criterion, returning all row pairs.
}
\examples{
ttb <- ttbModel(city_population, 3, c(4:ncol(city_population)))
reg <- regInterceptModel(city_population, 3, c(4:ncol(city_population)))
# which row pairs is ttb better on?
goal_type <- 'ProbGreater'
out <- aggregatePredictPair(list(ttb, reg), city_population, goal_type,
  rowIndexes())
out_df <- data.frame(out)
head(out_df[out_df$ttbModel > out_df$regInterceptModel,])
}

