% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{coef.singleCueModel}
\alias{coef.singleCueModel}
\title{Linear weights that can be used to compare pairs of cue directions.}
\usage{
\method{coef}{singleCueModel}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{...}{other arguments.}
}
\description{
Do NOT apply these directly to raw data.
}

