% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{coef.ttbModel}
\alias{coef.ttbModel}
\title{Linear weights that can be used to compare pairs of cue directions.}
\usage{
\method{coef}{ttbModel}(object, ...)
}
\arguments{
\item{object}{A fitted ttbModel.}
}
\description{
Do NOT apply these directly to raw data.
}

