% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{collapseConfusionMatrix3x3To2x2}
\alias{collapseConfusionMatrix3x3To2x2}
\title{Collapses a 3x3 confusion matrix to a 2x2 confusion matrix.}
\usage{
collapseConfusionMatrix3x3To2x2(
  confusion_matrix_3x3,
  guess_handling_fn = distributeGuessAsExpectedValue,
  tie_handling_fn = distributeTies
)
}
\arguments{
\item{confusion_matrix_3x3}{A 3x3 confusion matrix.}

\item{guess_handling_fn}{A function to call on the 3x3 confusion matrix to
assign a model's guesses-- 0 predictions tracked in the 2nd column-- to
-1 or 1 counts.}

\item{tie_handling_fn}{A function to call on the 3x3 confusion matrix to
distribute ties-- 0 correct answers tracked in the 2nd row-- to -1 or 1
counts.}
}
\value{
A 2x2 confusion matrix.
}
\description{
A 3x3 confusion matrix results from predictPair.
}
\details{
The middle column repressents guesses.
The middle row represents ties.  T
}
