% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cue_validity.R
\name{conditionalCueValidityMatrix}
\alias{conditionalCueValidityMatrix}
\title{Calculate conditional cue validity for cols_to_fit predicting criterion.}
\usage{
conditionalCueValidityMatrix(data, criterion_col, cols_to_fit)
}
\arguments{
\item{data}{The matrix or data.frame whose columns are treated as cues.}

\item{criterion_col}{The index of the column used as criterion.}

\item{cols_to_fit}{A vector of indexes of the columns to calculate cue
validity for.}
}
\value{
A vector of cue validities applying to each of the columns in
        cols_to_fit.  This *will* return Nan if the validity cannot
        be calculated (e.g. higher-validity cues make all the decisions).
        Your code must handle the NaN correctly.
}
\description{
Conditional cue validity is the validity of a cue taking into account
decisions already made by higher-validity cues.  For a single cue, it
is the same as cue validity.  For two cues, the higher validity cue will
have conditional cue validity = cue validity.  However, the remaining cue
will have its validity re-calculated on just those pairs of object where
cue validity did not discriminate.  In the case of binary data, there will
be many pairs where the first cue did not discriminate.  With real-valued
data, there may be no such cases.
}
\references{
Martignon, L., & Hoffrage, U.  (2002).  Fast, frugal, and fit: Simple
heuristics for paired comparisons.  Theory and Decision, 52: 29-71.
export
}
\seealso{
\code{\link{cueValidity}} for the unconditional version.
}

