% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{createErrorsFromPredicts}
\alias{createErrorsFromPredicts}
\title{Do not use.  Still under development.}
\usage{
createErrorsFromPredicts(df)
}
\arguments{
\item{df}{A dataframe to convert.  Assumes it has a column named correctProb
and predictions are in all but the first 3 columns.
(First 3 are Row1, Row2, and correct.)}
}
\value{
A dataframe with the last column converted to:
  0: if it matched correctProb
  1: if it was wrong.
  0.5: if it guessed or the true value was a guess and it wasn't a guess.
 This is technically a measure of error.
 TODO(jean): Make it work on all model columns, not just last column.
}
\description{
Do not use.  Still under development.
}

