% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{createFunction.heuristics}
\alias{createFunction.heuristics}
\title{Create function for heuristics prediction with rowPairApply.}
\usage{
\method{createFunction}{heuristics}(object, test_data)
}
\arguments{
\item{object}{A heuristics object.}

\item{test_data}{The test data that row_pairs will be drawn from.
We recommend}
}
\value{
A function that can easily be used by rowPairApply to
  generate predictions for all heuristics the object was created with.
  If it was created with M heuristics, it will generate M predictions.
}
\description{
Creates a function that takes an index pair and returns a prediction
for each of the predictProbInternal implementers.
}
\keyword{internal}
