% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{createFunction}
\alias{createFunction}
\title{Generic function to create functions for rowPairApply.}
\usage{
createFunction(object, test_data)
}
\arguments{
\item{object}{The object that implements createFunction, e.g. 
heuristicsProb(ttb).}

\item{test_data}{The test data that row_pairs will be drawn from.
We recommend}
}
\value{
A function that can easily be used by rowPairApply.  This
  is not normally used by ordinary users.
}
\description{
An example of solving something with another layer of indirection.
https://en.wikipedia.org/wiki/Fundamental_theorem_of_software_engineering
}
\keyword{internal}
