% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{createPctCorrectsFromErrors}
\alias{createPctCorrectsFromErrors}
\title{Do not use.  Still under development.}
\usage{
createPctCorrectsFromErrors(errors)
}
\arguments{
\item{errors}{A dataframe of heuristic errors to calculate with}
}
\value{
A dataframe with one row and the last column as a percent correct.
 TODO(jean): Make it work on all model columns, not just last column.
}
\description{
Do not use.  Still under development.
}

