% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{criterion}
\alias{criterion}
\title{Wrapper to get a column of correct probabilities with allRowPairApply.}
\usage{
criterion(criterion_col, output_column_name = "ProbGreater")
}
\arguments{
\item{criterion_col}{The integer index of the criterion in test_data.}

\item{output_column_name}{An optional string}
}
\value{
An object of class heuristics, which implements createFunction.
  Users will generally not use this directly-- allRowPairApply will.
}
\description{
Using allRowPairApply, this can generate a column with
the correct probability that row 1 > row 2 for each row pair in 
the test_data.  It can do this using the criterion column passed in.
By default, the output column is called ProbGreater, but you
can override the name with output_column_name.
}
\seealso{
\code{\link{createFunction}} which is what the returned object implements.

\code{\link{allRowPairApply}} which uses createFunction.
}

