% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cue_validity.R
\name{cueValidity}
\alias{cueValidity}
\title{Calculate the cue validity http://en.wikipedia.org/wiki/Cue_validity}
\usage{
cueValidity(criterion, cue, replaceNanWith = 0.5)
}
\arguments{
\item{criterion}{A vector of values to be predicted.}

\item{cue}{A vector of values to predict with.  Should have the same
length as the criterion.}

\item{replaceNanWith}{The value to return as cue validity in case it
cannot be calculated.}
}
\value{
The cue validity, a value in the range [0,1].
}
\description{
Calculate the cue validity http://en.wikipedia.org/wiki/Cue_validity
}

