% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{dawesModel}
\alias{dawesModel}
\title{DawesModel, a unit-weight linear model}
\usage{
dawesModel(train_data, criterion_col, cols_to_fit)
}
\arguments{
\item{train_data}{Training/fitting data as a matrix or data.frame.}

\item{criterion_col}{The index of the colum in train_data that has the criterion.}

\item{cols_to_fit}{A vector of column indices in train_data, used to fit the criterion.}
}
\value{
An object of \code{\link[base]{class}} dawesModel.  This is a list containing at least the following components:
  \itemize{
   \item "cue_validities": A list of cue validities for the cues in order of cols_to_fit.
   \item "linear_coef": A list of linear model coefficents (-1 or +1)
          for the cues in order of cols_to_fit.
  }
}
\description{
DawesModel is a unit-weight linear model inspired by Robyn Dawes.
Dawes Model assigns unit (+1 or -1) weights based on \code{\link{cueValidity}}.
  \itemize{
    \item A cue validity > 0.5 results in a weight of +1.
    \item A cue validity < 0.5 results in a weight of -1.
  }
This version differs from others in that it uses a weight of 0 if cue validity is 0.5
(rather than randomly assigning +1 or -1) to give faster convergence of average accuracy.
}
\seealso{
\code{\link{predict.dawesModel}} (via \code{\link[stats]{predict}}) for prediction.

Wikipedia's entry on \url{http://en.wikipedia.org/wiki/Unit-weighted_regression}.
}

