% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{getPredictiono}
\alias{getPredictiono}
\title{Extract prediction from output of predictPair for {row1, row2}.}
\usage{
getPredictiono(object, row1 = NULL, row2 = NULL)
}
\arguments{
\item{object}{a pairPredictor class}

\item{row1}{The index of row1 of the test_data that was predicted}

\item{row2}{The index of row2 of the test_data that was predicted}
}
\value{
The probability that row1 is greater
}
\description{
Extract prediction from output of predictPair for {row1, row2}.
}

