% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{getPredictionRowLC}
\alias{getPredictionRowLC}
\title{Assuming you have a matrix with a columns row1 and row2,
this helps you get the row which matches those columns.
getPrediction makes code below more readable.}
\usage{
getPredictionRowLC(df, row1 = NULL, row2 = NULL)
}
\arguments{
\item{df}{Data.frame to extra row from}

\item{row1}{The value in the row1 column to look for}

\item{row2}{The value in the row2 column to look for}
}
\value{
a row of the data frame.  (This could be multiple rows
  if multiple rows match.)
}
\description{
Assuming you have a matrix with a columns row1 and row2,
this helps you get the row which matches those columns.
getPrediction makes code below more readable.
}

