% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{heuristics}
\alias{heuristics}
\title{Wrap fitted heuristics to pass to rowPairApply to call predictPair.}
\usage{
heuristics(...)
}
\arguments{
\item{...}{A list of predictPairInternal implementers, e.g. a fitted ttb model.}
}
\value{
An object of class heuristics, which implements createFunction.
  Users will generally not use this directly-- rowPairApply will.
}
\description{
One or more fitted heuristics can be passed in.  They must all have the
same cols_to_fit.  If they differ on cols_to_fit, then group them in separate
heuristics functions.
}
\details{
Users will generally not use the output directly but instead pass this to
rowPairApply.
}
\examples{
## This is typical usage:
data <- cbind(y=c(30,20,10,5), x1=c(1,1,0,0), x2=c(1,1,0,1))
ttb <- ttbModel(data, 1, c(2:ncol(data)))
rowPairApply(data, heuristics(ttb))
## This outputs ttb's predictions for all 6 row pairs of data.
## (It has 6 row pairs because 4*2/2 = 6.)  It gets the predictions
## by calling ttb's predictPairInternal.

## It can handle multiple models at the same time.
unit <- unitWeightModel(data, 1, c(2:ncol(data)))
rowPairApply(data, heuristics(ttb, unit))
## This outputs predictions with columns 'ttbModel' and 'unitWeightModel'.

## But if another model uses different cols_to_fit, put it in a separate
## heuristics function.
ttb_just3 <- ttbModel(data, 1, c(3), fit_name="ttb_just3")
rowPairApply(data, heuristics(ttb, unit), heuristics(ttb_just3))
## This outputs predictions with columns 'ttbModel', 'unitWeightModel',
## and 'ttb_just3'.

}
\seealso{
\code{\link{heuristicsList}} for a version of this function where you can
control the function called (not necessarily predictPairInternal).

\code{\link{rowPairApply}} which is what heuristicsProb is passed in to.

\code{\link{predictPairInternal}} which must be implemented by heuristics in
   order to use them with the heuristics() wrapper function.
}

