% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{logAccuracy}
\alias{logAccuracy}
\title{Returns a correct proportion based on a vector of predictions.}
\usage{
logAccuracy(fit_predictions, test_data, criterion_col, cols_to_fit,
  row_pairs = NULL)
}
\arguments{
\item{fit_predictions}{A vector of predictions ranging from 0 to 1.}

\item{test_data}{test data.}

\item{criterion_col}{vector specifying criterion column}

\item{cols_to_fit}{vector specifying columns to fit}

\item{row_pairs}{Optional matrix.  TODO(jean): share documentation.}
}
\value{
Returns a single value ranging from 0 to 1.
private
}
\description{
Intended for logistic regression.
Assumes original data was sorted so that correct answers are all 1.
This may need to be temporarily exported but will eventually be private.
}

