% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{logRegSignModel}
\alias{logRegSignModel}
\title{Logistic Regression model using the sign of the difference of cues}
\usage{
logRegSignModel(train_data, criterion_col, cols_to_fit,
  cue_order_fn = keepOrder, suppress_warnings = TRUE)
}
\arguments{
\item{train_data}{Training/fitting data as a matrix or data.frame.}

\item{criterion_col}{The index of the colum in train_data that has the
criterion.}

\item{cols_to_fit}{A vector of column indices in train_data, used to fit
the criterion.}

\item{cue_order_fn}{Optional argument as a function that orders cues.  This
affects which cues are dropped for underspecified models. The rightmost
cues in the order are dropped first, so the function rankByCueValidity
means cues with the lowest cueValidity in the training set will be
be dropped first.  The function must have the signature
function(train_data, criterion_col, cols_to_fit).}

\item{suppress_warnings}{Optional argument specifying whether glm warnings
should be suppressed or not. Default is TRUE.}
}
\value{
An object of class logRegModel.
}
\description{
Create a logistic regression model by specifying columns and a dataset.
It fits the model with R's glm function.
}
\details{
This version assumes you do not want to include the intercept.
}

