% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cue_validity.R
\name{matrixCueValidity}
\alias{matrixCueValidity}
\title{Calculate the cue validity for every column in the matrix.}
\usage{
matrixCueValidity(matrix, criterionColIndex, replaceNanWith = 0.5,
  includeCriterionValidity = FALSE)
}
\arguments{
\item{matrix}{The matrix whose columns are treated as cues.}

\item{criterionColIndex}{The index of the column used as criterion.}

\item{replaceNanWith}{See cueValidity.}

\item{includeCriterionValidity}{Whether to return the cue validity of
the criterion, which will always be 1.}
}
\description{
Calculate the cue validity for every column in the matrix.
}

