% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{minModel}
\alias{minModel}
\title{Minimalist Model}
\usage{
minModel(train_data, criterion_col, cols_to_fit, reverse_cues = FALSE)
}
\arguments{
\item{train_data}{Training/fitting data as a matrix or data.frame.}

\item{criterion_col}{The index of the colum in train_data that has the criterion.}

\item{cols_to_fit}{A vector of column indices in train_data, used to fit the criterion.}

\item{reverse_cues}{Optional parameter to reverse cues as needed.}
}
\value{
An object of class minModel.
}
\description{
Fit the Minimalist heuristic by specifying columns and a dataset.
}
\seealso{
\code{\link{predictPair.minModel}} (via \code{\link{predictPair}}) for prediction.


}

