% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{oneRow}
\alias{oneRow}
\title{Convenience function to get one row from a matrix or data frame.}
\usage{
oneRow(matrix_or_data_frame, row_index)
}
\arguments{
\item{matrix_or_data_frame}{A matrix or data frame from which you want one
row.}

\item{row_index}{The integer index of the row}
}
\value{
The selected row of the data frame.
}
\description{
This simply calls matrix_or_data_frame[row_index,,drop=FALSE] for you
but is shorter and helps you avoid forgetting drop=FALSE.  The need
for drop=FALSE when selecting just one row is explained here:
http://www.hep.by/gnu/r-patched/r-faq/R-FAQ_56.html
}
