% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{pairMatrix}
\alias{pairMatrix}
\title{Apply a function to all unique pairs of row indices up to num_row.}
\usage{
pairMatrix(num_row, pair_evaluator_fn)
}
\arguments{
\item{num_row}{The number of rows to generate index pairs for.}

\item{pair_evaluator_fn}{The function you want applied.  It should
accept a list of two numbers, the index of row 1 and the index of row2.}
}
\value{
A matrix of the output of the function for all unique row pairs:
   c(pair_evaluator_fn(c(1,2), pair_evaluator_fn(c(1,3)), etc.)
}
\description{
Apply a function to all unique pairs of row indices up to num_row.
}

