% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{pctCorrectOfPredictAlternative}
\alias{pctCorrectOfPredictAlternative}
\title{Do not use.  Still under development.}
\usage{
pctCorrectOfPredictAlternative(fitted_heuristic_list, test_data)
}
\arguments{
\item{fitted_heuristic_list}{A list of heuristics already fitted to data, e.g. ttbBinModel.}

\item{test_data}{Data to try to predict; must match columns in fit.}
}
\value{
A numeric from 0 to 1 indicating the percent correct.
}
\description{
Do not use.  Still under development.
}

