% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{percentCorrectListReturnMatrix}
\alias{percentCorrectListReturnMatrix}
\title{Percent correct of heuristics' predictPair on test_data, returning a matrix.}
\usage{
percentCorrectListReturnMatrix(test_data, fitted_heuristic_list)
}
\arguments{
\item{test_data}{Data to try to predict.  Must have same criterion column
and cols_to_fit as the data heuristics were fit to.}

\item{fitted_heuristic_list}{A list of one or more heuristics fitted to
data, e.g. the output of ttbModel.}
}
\value{
A one-row matrix of numbers from 0 to 100, the percent correct
  of each heuristic.  Each column is named with the heuristic's class or
  the fit name.
}
\description{
Percent correct of heuristics' predictPair on test_data, returning a matrix.
}
\examples{
# See examples for percentCorrectList, which returns a data.frame.

}
\seealso{
\code{\link{percentCorrectList}} for a version that returns a
  data.frame and includes several examples.
}
