% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{percentCorrectNonSymmetric}
\alias{percentCorrectNonSymmetric}
\title{percentCorrect for non-symmetric heuristics}
\usage{
percentCorrectNonSymmetric(fitted_heuristic_list, test_data)
}
\arguments{
\item{fitted_heuristic_list}{A list of heuristics already fitted to data,
e.g. ttbModel.}

\item{test_data}{Data to try to predict; must match columns in fit.}
}
\value{
A one-row matrix of numbers from 0 to 1, meaning proportion
  correct.  Each column is named with the heuristic's class.
}
\description{
Same as percentCorrect but for weird heuristics that do not
consistently choose the same row.  When a symmetric heuristic predicts
row1 > row2, then it also predicts row2 < row1.  Those can be used
with percentCorrect.  All heuristics built into heuristica
qualify.  They will get the same answers for percentCorrect
and percentCorrectNonSymmetric.  But a non-symmetric heuristic
will only get correct answers for percentCorrectNonSymmetric.
}
\seealso{
\code{\link{percentCorrect}} for prediction.
}

