% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predict.franklinModel}
\alias{predict.franklinModel}
\title{Generates predictions for Franklin's Model}
\usage{
\method{predict}{franklinModel}(object, ...)
}
\arguments{
\item{object}{A fitted franklinModel.}

\item{...}{Normally this would be the test data. 
It is used to predict and can be a matrix or data.frame.  
It must have the same cols_to_fit indices as those used in train_data.}
}
\value{
An N x 1 matrix of predicted values, or a list if there was only one cue.
}
\description{
Implementation of \code{\link[stats]{predict}} for franklinModel.
}

