% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predict.ttbBinModel}
\alias{predict.ttbBinModel}
\title{Generates predictions for Take The Best with binary cues}
\usage{
\method{predict}{ttbBinModel}(object, ...)
}
\arguments{
\item{object}{A ttbBinModel.}

\item{...}{Normally this would be the test data.
 It is used to predict and can be a matrix or data.frame.
 It must have the same cols_to_fit indices as those used in train_data.}
}
\value{
An N x 1 matrix of predicted values, or a list if there was only one cue.
 Only the sort order of these is relevant for Take The Best.
}
\description{
Implementation of \code{\link[stats]{predict}} for ttbBinModel,
Take The Best with binary cues.
}
\seealso{
\code{\link{ttbBinModel}} for example code.
}

