% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predict.ttbModel}
\alias{predict.ttbModel}
\title{Generates predictions for Take The Best}
\usage{
\method{predict}{ttbModel}(model, test_df)
}
\description{
Implementation of \code{\link[stats]{predict}} for ttbModel.
}

