% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictAlternative.franklinModel}
\alias{predictAlternative.franklinModel}
\title{Predict which alternative has higher criterion for Franklin model.}
\usage{
\method{predictAlternative}{franklinModel}(object, test_data, rowPairs = NULL)
}
\arguments{
\item{object}{A fitted franklinModel.}

\item{test_data}{The matrix of data to predict on.  As with predict, columns
must match those used for fitting.}

\item{rowPairs}{The optionalrowPairs is a matrix of pairs of column indices
you want to have predicted.  By default (if no list is given), it will use
all unique pairs.
E.g. for 3 rows, it will use [[1,2], [1,3], [2,3]].}
}
\description{
Predict which alternative has higher criterion for Franklin model.
}
\seealso{
\code{\link{franklinModel}} for example code.
}

