% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictAlternative.singleCueModel}
\alias{predictAlternative.singleCueModel}
\title{Predict which alternative has higher criterion for Single Cue Model.}
\usage{
\method{predictAlternative}{singleCueModel}(object, test_data,
  row_pairs = NULL)
}
\arguments{
\item{object}{A singleCueModel.}

\item{test_data}{The matrix of data to predict on.  As with predict, columns
must match those used for fitting.}

\item{row_pairs}{The optionalrow_pairs is a matrix of pairs of column indices
you want to have predicted.  By default (if no list is given), it will use
all unique pairs.
E.g. for 3 rows, it will use [[1,2], [1,3], [2,3]].}
}
\description{
Predict which alternative has higher criterion for Single Cue Model.
}
\seealso{
\code{\link{singleCueModel}} for example code.
}

