% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictAlternative.ttbBinModel}
\alias{predictAlternative.ttbBinModel}
\title{Predict which alternative has higher criterion for Take The Best with binary cues}
\usage{
\method{predictAlternative}{ttbBinModel}(object, ...)
}
\arguments{
\item{object}{A ttbBinModel.}

\item{...}{Normally this would be the test data.
 It is used to predict and can be a matrix or data.frame.
 It must have the same cols_to_fit indices as those used in train_data.
TODO: User can pass in i, j for the rwos to compare, defaulting to 1 and 2.}
}
\value{
1 if the first row is predicted greater,
 -1 if the 2nd row is predicted greater,
 0 if the rows are predicted to have equal criterion values.
}
\description{
Predict which alternative has higher criterion for Take The Best with binary cues
}
\seealso{
\code{\link{ttbBinModel}} for example code.
}

