% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictAlternativeWithCorrect}
\alias{predictAlternativeWithCorrect}
\title{Do not use.  Still under development.}
\usage{
predictAlternativeWithCorrect(fitted_heuristic, test_data)
}
\arguments{
\item{fitted_heuristic}{A heuristic already fitted to data, e.g. ttbBinModel.}

\item{test_data}{Data to try to predict; must match columns in fit.}
}
\value{
Same matrix as predictAlternative but with columns on correctness
}
\description{
Do not use.  Still under development.
}
\seealso{
\code{\link{predictAlternative}}
}

