% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_with_weights.R
\name{predictAlternativeWithWeights}
\alias{predictAlternativeWithWeights}
\title{Predict alternative for each pair of row, assigning weights to each row.}
\usage{
predictAlternativeWithWeights(test_data, cols_to_fit, col_weights,
  row_pairs = NULL)
}
\arguments{
\item{test_data}{Data to predict for, as either a matrix or a data.frame.}

\item{cols_to_fit}{Vector of column indexes to use in test_data.}

\item{col_weights}{Vector of weights to apply to the columns indicated by cols_to_fit.}

\item{row_pairs}{Optional matrix.  TODO(jean): share documentation.}
}
\value{
A data.frame (rows * (rows-1)) of predictions for each possible paired comparison.
  Description of each column:
  Row1: The index of row1 of the comparison.
  Row2: The index of row2 of the comparison.
  predictDirection (Prob_Row1_Bigger): The prediction of the probability that row1 has the bigger criterion.
    Specifically, 1 means the first row is bigger, 0 means the 2nd row, 0.5 is a guess.
}
\description{
Applies the weights in col_weights to the columns cols_to_fit in test_data, then
uses the row-level predictions to predict alternatives.
}

