% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictPair.logRegWithIModel}
\alias{predictPair.logRegWithIModel}
\title{Predict which of a pair of rows has a higher criterion, using logistic regression.}
\usage{
\method{predictPair}{logRegWithIModel}(object, test_data,
  verbose_output = TRUE)
}
\arguments{
\item{object}{A fitted logRegModel.}

\item{test_data}{The matrix of data to predict on.  As with predict, columns
must match those used for fitting.}

\item{verbose_output}{Controls how much output is generated, which may slow down
computations and use more memory.  When automating, set to FALSE, which will turn
off outputs preced by verbose_.}
}
\description{
Predict which of a pair of rows has a higher criterion, using logistic regression.
}
\seealso{
\code{\link{logRegWithIModel}} for example code.
}

