% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictPair.minModel}
\alias{predictPair.minModel}
\title{Predict which of a pair of rows has a higher criterion, using Minimalist Model.}
\usage{
\method{predictPair}{minModel}(object, test_data, verbose_output = TRUE)
}
\arguments{
\item{object}{A fitted minModel.}

\item{test_data}{The matrix of data to predict on.  As with predict, columns
must match those used for fitting.}

\item{verbose_output}{Controls how much output is generated, which may slow down
computations and use more memory.  When automating, set to FALSE, which will turn
off outputs preced by verbose_.}
}
\description{
Predict which of a pair of rows has a higher criterion, using Minimalist Model.
}
\seealso{
\code{\link{minModel}} for example code.
}

