% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictPair}
\alias{predictPair}
\title{Generic function to predict which of a pair of rows has a higher criterion.
EXPERIMENTAL and not meant for general use yet.}
\usage{
predictPair(object, test_data, subset = NULL)
}
\arguments{
\item{object}{The object that implements predictAlternative, e.g. a ttb model.}

\item{test_data}{The matrix of data to predict on.  As with predict, columns
must match those used for fitting.}

\item{subset}{Vector of row indices-- all pairs of these will be predicted.}
}
\value{
A vector of probabilities that the first row has a greater criterion.
}
\description{
Generic function to predict which of a pair of rows has a higher criterion.
EXPERIMENTAL and not meant for general use yet.
}

