% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictPair.singleCueModel}
\alias{predictPair.singleCueModel}
\title{Predict which of a pair of rows has a higher criterion, using Single Cue Model.}
\usage{
\method{predictPair}{singleCueModel}(object, test_data, subset_rows = NULL,
  verbose_output = TRUE)
}
\arguments{
\item{object}{A fitted singleCueModel.}

\item{test_data}{The matrix of data to predict on.  As with predict, columns
must match those used for fitting.}

\item{subset_rows}{Vector of row indices-- all pairs of these will be predicted.}

\item{verbose_output}{Controls how much output is generated, which may slow down
computations and use more memory.  When automating, set to FALSE, which will turn
off outputs preced by verbose_.}
}
\description{
Predict which of a pair of rows has a higher criterion, using Single Cue Model.
}
\seealso{
\code{\link{singleCueModel}} for example code.
}

