% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictPair.ttbModel}
\alias{predictPair.ttbModel}
\title{Predict which of a pair of rows has higher criterion, using Take The Best.
This is a lightning-fast, non-debuggable version of predictAlternative.}
\usage{
\method{predictPair}{ttbModel}(object, test_data, subset = NULL)
}
\arguments{
\item{object}{A fitted ttbModel.}

\item{test_data}{The matrix of data to predict on.  As with predict, columns
must match those used for fitting.}

\item{subset}{Vector of row indices-- all pairs of these will be predicted.}
}
\description{
Predict which of a pair of rows has higher criterion, using Take The Best.
This is a lightning-fast, non-debuggable version of predictAlternative.
}
\seealso{
\code{\link{ttbModel}} for example code.
}

