% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{predictPairConfusionMatrix}
\alias{predictPairConfusionMatrix}
\title{Make a confusion matrix using predictPair with the heuristic and test_data.}
\usage{
predictPairConfusionMatrix(test_data, fitted_heuristic,
  guess_handling_fn = distributeGuessAsExpectedValue,
  tie_handling_fn = distributeTies, symmetric_model = TRUE)
}
\arguments{
\item{test_data}{A data set.  Must have the criterion_column and cols_to_fit
that the heuristic expects.}

\item{fitted_heuristic}{A fitted heuristic that implements predictPair}

\item{guess_handling_fn}{A function to call on the 3x3 confusion matrix to
assign a model's guesses-- 0 predictions tracked in the 2nd column-- to
-1 or 1 counts.}

\item{tie_handling_fn}{A function to call on the 3x3 confusion matrix to
distribute ties-- 0 correct answers tracked in the 2nd row-- to -1 or 1
counts.}

\item{symmetric_model}{Optional parameter that is TRUE by default because
all models in heuristica are symmtric.  (For an asymmteric model, this
function will run both A vs. B and B vs. A through predicPair.)}
}
\value{
A 2x2 confusion matrix with rows and columns for -1 and 1.  It will
  be 2x2 even if predictions and correct do not cover the full range.
}
\description{
Make a confusion matrix using predictPair with the heuristic and test_data.
}
\references{
Wikipedia's entry on
\url{https://en.wikipedia.org/wiki/Confusion_matrix}.
}

