% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictPairMatrix}
\alias{predictPairMatrix}
\title{Run predictPair for all row pairs in test_data to generate a one-column matrix.}
\usage{
predictPairMatrix(object, test_data)
}
\arguments{
\item{object}{A fitted model that implements predictRoot.}

\item{test_data}{A matrix of data frame or data with object$cols_to_fit columns.}
}
\value{
An one-column matrix.  Each row represents a pair of rows.  Use get
  prediction functions to find the rows if you care.  (TODO(jean): Document that
  more-- give examples, too.)
}
\description{
Run predictPair for all row pairs in test_data to generate a one-column matrix.
}
\seealso{
\code{\link{predictRoot}} for the function the fitted model must implement.
}

