% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{predictPairProb}
\alias{predictPairProb}
\title{Predict the probablity that row1 has a higher criterion than row2.}
\usage{
predictPairProb(row1, row2, object)
}
\arguments{
\item{row1}{The first row of cues (will apply cols_to_fit for you, based on
object).}

\item{row2}{The second row (will apply cols_to_fit for you, based on
object).}

\item{object}{The object that implements predictPair, e.g. a fitted ttbModel
or regModel.}
}
\value{
A double from 0 to 1, representing the probability that row1's
  criterion is greater than row2's criterion.
}
\description{
It uses the passed-in object to do the prediction, assuming the object
implements predictRoot and has $cols_to_fit.
}

