% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{predictPairWithCorrect}
\alias{predictPairWithCorrect}
\title{Generates a matrix of all predictRowPair predictions plus correct output.}
\usage{
predictPairWithCorrect(fitted_heuristic_list, test_data)
}
\arguments{
\item{fitted_heuristic_list}{List of heuristics that implement the generic function
predictPair, e.g. ttbModel.  All heuristics must agree on the criterion_col.}

\item{test_data}{Data to try to predict; must match columns in fit.}
}
\value{
Same matrix as predictAlternative but with columns on correctness
}
\description{
This geneartes a column of correct output (whether row 1 or row 2 is greater) from
the test matrix, then runs all the heuristics in order, generating a column of
predictions for each, naming each column from the heuristic class.
}
\seealso{
\code{\link{predictRowPair}}
}

