% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_with_weights.R
\name{predictWithWeightsLog}
\alias{predictWithWeightsLog}
\title{Logistic prediction for use by logistic regression.}
\usage{
predictWithWeightsLog(test_data, cols_to_fit, criterion_col, col_weights)
}
\arguments{
\item{test_data}{Data to predict for, as either a matrix or a data.frame.}

\item{cols_to_fit}{Vector of column indexes to use in test_data.}

\item{criterion_col}{Column index specifying the criterion.}

\item{col_weights}{Vector of weights to apply to the columns indicated by cols_to_fit.}
}
\value{
A matrix (rows x1) of predictions, or a list if there was just one
   column to fit.

Special features:
Treats a weight of NA as zero (useful for rank-deficient regression fits).
If there is a column named "(Intercept)" in col_weights, it is added.
In other words, there is no need to add an intercept column to test_matrix.
}
\description{
Applies the weights in col_weights to the columns cols_to_fit in test_data.
}

