% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_with_weights.R
\name{predictWithWeightsMin}
\alias{predictWithWeightsMin}
\title{Linear prediction for use by minModel.}
\usage{
predictWithWeightsMin(test_data, cols_to_fit, col_weights)
}
\arguments{
\item{test_data}{Data to predict for, as either a matrix or a data.frame.}

\item{cols_to_fit}{Vector of column indexes to use in test_data.}

\item{col_weights}{Vector of weights to apply to the columns indicated by cols_to_fit.}
}
\value{
A matrix (rows x1) of predictions, or a list if there was just one
   column to fit.
}
\description{
Randomly shuffles weights in col_weights and applies them to the columns cols_to_fit in test_data.
}

