% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{regModelForDocumentation}
\alias{regModelForDocumentation}
\title{Just to share documentation}
\usage{
regModelForDocumentation(train_matrix, criterion_col, cols_to_fit)
}
\arguments{
\item{train_matrix}{A matrix (or data.frame) of data to train (fit) the model with.}

\item{criterion_col}{The index of the criterion column-- “y” in the formula.}

\item{cols_to_fit}{A vector of column indexes to fit-- the “x’s” in the formula.}
}
\description{
Just to share documentation
}

