% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{reverseRowsAndColumns}
\alias{reverseRowsAndColumns}
\title{Reverse rows and columns of data}
\usage{
reverseRowsAndColumns(data)
}
\arguments{
\item{data}{A data.frame or matrix.}
}
\value{
A data.frame or matrix with rows reversed and columns reversed.
}
\description{
This matrix
     [,1] [,2]
[1,]    1    2
[2,]    3    4
becomes
      [,1] [,2]
[1,]    4    3
[2,]    2    1
}

