% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{rowIndexes}
\alias{rowIndexes}
\title{Wrapper to output two columns, row 1 and row 2.}
\usage{
rowIndexes(rowIndexColNames = c("Row1", "Row2"))
}
\arguments{
\item{rowIndexColNames}{An optional vector of 2 strings for column names.}
}
\value{
An object of class rowIndexes, which implements createFunction.
  Users will generally not use this directly-- rowPairApply will.
}
\description{
Using rowPairApply, this can generate two columns, which by default
are called "Row1" and "Row2"
}
\seealso{
\code{\link{createFunction}} which is what the returned object implements.

\code{\link{rowPairApply}} which uses createFunction.
}
