% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{rowPairApply}
\alias{rowPairApply}
\title{Apply all functions to the row pair.}
\usage{
rowPairApply(row1, row2, ...)
}
\arguments{
\item{row1}{The first row of cues (will apply cols_to_fit for you, based
on object).}

\item{row2}{The second row (will apply cols_to_fit for you, based on
object).}

\item{...}{The objects that generate the functions to apply, using
createFunction.  For example, heuristics(ttb), probGreater(col), or
colPairValues.}
}
\value{
A matrix of function outputs.
}
\description{
Apply all functions to the row pair.
}

