% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{rowPairApply2Rows}
\alias{rowPairApply2Rows}
\title{Apply all functions to the two rows passed in.}
\usage{
rowPairApply2Rows(row1, row2, ...)
}
\arguments{
\item{row1}{The first row of cues (will apply cols_to_fit for you, based
on object).}

\item{row2}{The second row (will apply cols_to_fit for you, based on
object).}

\item{...}{The functions that generate the functions to apply, such as
heuristics(ttb) and correctGreater(col).}
}
\value{
A matrix of function outputs.
}
\description{
This has some asserts that exactly one row is passed in and exaclty one row
is returned, but otherwise it just calls rowPairApply.
}
\seealso{
\code{\link{rowPairApply}} to apply to all row pairs in a matrix or
  data.frame.
}
\keyword{internal}
