% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{rowPairApplyList}
\alias{rowPairApplyList}
\title{Apply functions to all row pairs.}
\usage{
rowPairApplyList(test_data, function_creator_list)
}
\arguments{
\item{test_data}{The data to apply the functions to as a matrix or
data.frame.  Heuristics must have already been fitted to trying data and
must include the same criterion_col and cols_to_fit.}

\item{function_creator_list}{List of the functions that generate the
functions to apply, such as heuristics(ttb) and correctGreater(col).}
}
\value{
A matrix of outputs from the functions.  The number of rows is based
  on the number of row pairs in test_data.  If the input has N rows, the
  output will have N x (N-1) rows.  The number of columns will be at least
  the number of functions but may be more as some functions may output more
  than one column.
}
\description{
Apply functions like heuristic predictions to all row pairs in a matrix
or data.frame.
}
\examples{
# This function is called like 
# rowPairApplyList(data, list(heuristics(ttb, reg)))
# instead of
# rowPairApply(data, heuristics(ttb, reg))
# See rowPairApply for details.

}
\seealso{
\code{\link{rowPairApply}} for no need to use a list.  Examples and details
  are there.
}

