% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_with_weights.R
\name{rowPairGenerator}
\alias{rowPairGenerator}
\title{Generates a matrix of pairs of row indices for two-alternative choice.}
\usage{
rowPairGenerator(n)
}
\arguments{
\item{n}{is the number of rows.}
}
\value{
Returns a data.frame with (n x n-1) rows and 2 columns.  The column
names are Row1 and Row2.  Rows are sorted by Row1, then Row2.
}
\description{
Generates all pairs, including repeating pairs in reverse order, but not a
row with itself.  Output is sorted by Row1, then Row2.
}
\examples{
rowPairGenerator(2)
# You should get:
#  Row1 Row2
#1    1    2
#2    2    1

rowPairGenerator(3)
# You should get:
#  Row1 Row2
#1    1    2
#2    1    3
#3    2    1
#4    2    3
#5    3    1
#6    3    2

}

