% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{ttbModel}
\alias{ttbModel}
\title{Take The Best}
\usage{
ttbModel(train_data, criterion_col, cols_to_fit)
}
\arguments{
\item{train_data}{Training/fitting data as a matrix or data.frame.}

\item{criterion_col}{The index of the colum in train_data that has the criterion.}

\item{cols_to_fit}{A vector of column indices in train_data.}
}
\value{
The ttb S3 object.
}
\description{
An S3 class implementation of the Take The Best heuristic.
It sorts cues in order of cue validity, making a decision based on the first cue that
discriminates (has differing values on the two objects).
Implements:
  predict
However, its implementation of the S3 predict generic function has to output values.
Do not take the actual values seriously-- only the sort order is guaranteed.
}
\examples{
# Fit column (5,4) to column (1,0), having validity 1.0, and column (0,1), validity 0.
ttb <- ttbModel(matrix(c(5,4,1,0,0,1), 2, 3), 1, c(2,3))
predict(ttb, matrix(c(5,4,1,0,0,1), 2, 3))  # Only sort order of output guaranteed.
}

