% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{zzDocumentationStubFormulaModelParams}
\alias{zzDocumentationStubFormulaModelParams}
\title{Documentation stub. Just to share documentation.}
\usage{
zzDocumentationStubFormulaModelParams(train_matrix, criterion_col, cols_to_fit)
}
\arguments{
\item{train_matrix}{A matrix (or data.frame) of data to train (fit) the
model with.}

\item{criterion_col}{The index of the criterion column-- "y" in the formula.}

\item{cols_to_fit}{A vector of column indexes to fit-- the "x's" in the
formula.}
}
\description{
Documentation stub. Just to share documentation.
}
\keyword{internal}
