% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics.R
\name{zzDocumentationStubModelParams}
\alias{zzDocumentationStubModelParams}
\title{Documentation stub.  Just a way to share parameter documentation.}
\usage{
zzDocumentationStubModelParams(train_data, criterion_col, cols_to_fit)
}
\arguments{
\item{train_data}{Training/fitting data as a matrix or data.frame.}

\item{criterion_col}{The index of the colum in train_data that has the
criterion.}

\item{cols_to_fit}{A vector of column indices in train_data, used to fit
the criterion.}
}
\description{
Documentation stub.  Just a way to share parameter documentation.
}
\keyword{internal}

