% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cue_validity.R
\name{cueValidityMatrix}
\alias{cueValidityMatrix}
\title{Calculate the cue validity for cols_to_fit predicting the criterion.}
\usage{
cueValidityMatrix(data, criterion_col, cols_to_fit, replaceNanWith = 0.5,
  reverse_cues = FALSE)
}
\arguments{
\item{data}{The matrix or data.frame whose columns are treated as cues.}

\item{criterion_col}{The index of the column used as criterion.}

\item{cols_to_fit}{A vector of indexes of the columns to calculate cue
validity for.}

\item{replaceNanWith}{The value to return as cue validity in case it
cannot be calculated.}

\item{reverse_cues}{Optional parameter to reverse cues as needed.  By
default, the model will reverse the cue values for cues with cue validity
< 0.5, so a cue with validity 0 becomes a cue with validity 1.
Set this to FALSE if you do not want that, i.e. the cue stays validity 0.}
}
\value{
A list where $cue_validities has a vector of validities for
  each of the columns in cols_to_fit.
}
\description{
If you know you want cue validities for many columns in your data,
this function makes it easy.
}
\references{
Wikipedia's entry on
\url{https://en.wikipedia.org/wiki/Cue_validity}
}

