% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{franklinModel}
\alias{franklinModel}
\title{Franklin's Model, a linear model weighted by cue validities}
\usage{
franklinModel(train_data, criterion_col, cols_to_fit, reverse_cues = TRUE)
}
\arguments{
\item{train_data}{Training/fitting data as a matrix or data.frame.}

\item{criterion_col}{The index of the colum in train_data that has the criterion.}

\item{cols_to_fit}{A vector of column indices in train_data, used to fit the criterion.}

\item{reverse_cues}{Optional parameter to reverse cues as needed.  By default,
the model will reverse the cue values for cues with cue validity < 0.5, so a cue
with validity 0 becomes a cue with validity 1.
Set this to FALSE if you do not want that, i.e. the cue stays validity 0.}
}
\value{
An object of \code{\link[base]{class}} franklinModel.  This is a list containing at least the following components:
  \itemize{
   \item "cue_validities": A list of cue validities for the cues in order of cols_to_fit.
   \item "linear_coef": Same as cue validities for this model.
  }
}
\description{
Franklin's Model is a linear model with weights calculated by \code{\link{cueValidity}}.
The name is because it was inspired by a method used by Ben Franklin.
}
\seealso{
\code{\link{predict.franklinModel}} (via \code{\link[stats]{predict}}) for prediction.

\code{\link{predictAlternative.franklinModel}} for predicting among alternatives.
}

