% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{heuristics}
\alias{heuristics}
\title{Wrapper for fitted heuristics to generate predictions with allRowPairApply.}
\usage{
heuristics(...)
}
\arguments{
\item{...}{A list of predictRoot implementers, e.g. a fitted ttb model.}
}
\value{
An object of class heuristics, which implements createFunction.
  Users will generally not use this directly-- allRowPairApply will.
}
\description{
One or more fitted heuristics can be passed in.  They must all implement
predictRoot.  Users will generally not use the output directly.
}
\examples{
## This is typical usage:
ttb <- ttbModel(city_population, 3, c(4:ncol(city_population)))
out <- allRowPairApply(city_population, heuristics(ttb))
head(out)

## If you are curious how the function_creator works, try this:
function_creator <- heuristics(ttb)
fn <- createFunction(function_creator, city_population)
## ttb's prediction for row 1 vs. 2 (probability row 1 greater):
fn(c(1,2))
## ttb's prediction for row 2 vs. 1 (probability row 2 greater):
fn(c(1,2))
## ttb's prediction for row 82 vs 83 (probability row 82 greater):
fn(c(1,2))
## Under the hood, it calls ttb's predictRoot.

}
\seealso{
\code{\link{predictRowPair}} for a simpler way to generate predictions
  from fitted models.

\code{\link{predictRoot}} which must be implemented by heuristics in
   order to use them with the heuristics() wrapper function.

\code{\link{createFunction}} which is what the returned object implements.

\code{\link{allRowPairApply}} which uses createFunction.
}

