% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{heuristicsList}
\alias{heuristicsList}
\title{Wrapper for fitted heuristics to generate predictions with allRowPairApply.}
\usage{
heuristicsList(list_of_fitted_heuristics)
}
\arguments{
\item{list_of_fitted_heuristics}{A list of predictRoot implementers,
e.g. a fitted ttb model.}
}
\value{
An object of class heuristics, which implements createFunction.
  Users will generally not use this directly-- allRowPairApply will.
}
\description{
One or more fitted heuristics can be passed in.  They must all implement
predictRoot.  Users will generally not use the output directly.
}
\examples{
## This is typical usage:
ttb <- ttbModel(city_population, 3, c(4:ncol(city_population)))
out <- allRowPairApply(city_population, heuristics(ttb))
head(out)
## Under the hood, it calls ttb's predictRoot.

}
\seealso{
\code{\link{predictRowPair}} for a simpler way to generate predictions
  from fitted models.
\code{\link{heuristics}} for a non-list version of this function.

\code{\link{predictRoot}} which must be implemented by heuristics in
   order to use them with the heuristics() wrapper function.

\code{\link{createFunction}} which is what the returned object implements.

\code{\link{allRowPairApply}} which uses createFunction.
}

