% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{logAccuracy}
\alias{logAccuracy}
\title{Returns a correct proportion based on a vector of predictions.}
\usage{
logAccuracy(fit_predictions)
}
\arguments{
\item{vector}{of predictions ranging from 0 to 1.}
}
\value{
Returns a single value ranging from 0 to 1.
private
}
\description{
Intended for logistic regression.
Assumes original data was sorted so that correct answers are all 1.
This may need to be temporarily exported but will eventually be private.
}

