% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cue_validity.R
\name{matrixCueValidity}
\alias{matrixCueValidity}
\title{Calculate the cue validity for all specified columns}
\usage{
matrixCueValidity(data, criterion_col, cols_to_fit, replaceNanWith = 0.5)
}
\arguments{
\item{data}{The matrix or data.frame whose columns are treated as cues.}

\item{criterion_col}{The index of the column used as criterion.}

\item{cols_to_fit}{A vector of indexes of the columns to calculate cue
validity for}

\item{replaceNanWith}{The value to return as cue validity in case it
cannot be calculated.}
}
\description{
If you know you want cue validities for many columns in your data,
this function makes it easy.
}

