% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregators.R
\name{pctCorrectOfPredictAlternative}
\alias{pctCorrectOfPredictAlternative}
\title{Runs predictAlternative for fitted heuristics to get percent correct on test.}
\usage{
pctCorrectOfPredictAlternative(fitted_heuristic_list, test_data)
}
\arguments{
\item{fitted_heuristic_list}{A list of heuristics already fitted to data, e.g. ttbBinModel.}

\item{test_data}{Data to try to predict; must match columns in fit.}
}
\value{
A one-row matrix of numbers from 0 to 1 indicating the percent correct.
}
\description{
Combines other functons to
1. Create predictions of alternatives for all the heuristics in the list.
2. Get errors of those predictions.
3. Calculate overall percent correct for each heuristic.
Assumes the heuristics passed in have already been fitted to train_data.
}
\details{
Each of the 3 steps above can be output if you call the helper functions directly:
1. predictAlternativeWithCorrect
2. createErrorsFromPredicts
3. createPctCorrectsFromErrors
}

