% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregators.R
\name{pctCorrectOfPredictPair}
\alias{pctCorrectOfPredictPair}
\title{Percent correct of heuristics' predictPair on test_data.}
\usage{
pctCorrectOfPredictPair(fitted_heuristic_list, test_data)
}
\arguments{
\item{fitted_heuristic_list}{A list of heuristics already fitted to data,
e.g. ttbModel.}

\item{test_data}{Data to try to predict; must match columns in fit.}
}
\value{
A one-row data.frame of numbers from 0 to 1, meaning proportion
  correct.  Each column is named with the heuristic's class or the fit name.
}
\description{
Returns overall percent correct for all heuristics.
1. Create predictions using predictPair for all row pairs for all
fitted heuristics in the list.
2. Calculate percent correct for each heuristic.
Assumes the heuristics passed in have already been fitted to training
data and all have the same criterion column.
}
\examples{
df <- data.frame(y=c(30,20,10,5), name=c("a", "b", "c", "d"),
                 x1=c(1,1,0,0), x2=c(1,1,0,1))
ttb <- ttbModel(df, 1, c(3:4))
sing <- singleCueModel(df, 1, c(3:4))
pctCorrectOfPredictPair(list(ttb, sing), df)
#    ttbModel singleCueModel
#  1     0.75      0.8333333
# TTB gets 75\% correct while single cue model gets 83\%.

# Now repeatedly sample 2 rows of the data set and see how outcomes are
# affected, tracking with the fit_name.
set.seed(1) # If you want to reproduce the same output as below.

}
\seealso{
using one fitted heuristic.
}

