% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictAlternative}
\alias{predictAlternative}
\title{Generic function to use two sets of cues to predict which set has a higher criterion.}
\usage{
predictAlternative(x)
}
\description{
Generic function to use two sets of cues to predict which set has a higher criterion.
}

