% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictAlternative}
\alias{predictAlternative}
\title{Generic function to use two sets of cues to predict which set has a higher criterion.}
\usage{
predictAlternative(object, ...)
}
\arguments{
\item{object}{The object that implements predictAlternative, e.g. a ttb model.}

\item{...}{The 2nd optional parameter is a matrix of pairs of column indices
 you want to have predicted.  If no list is given, it will use all pairs.
E.g. for 3 rows, it will use [[1,2],   [1,3],   [2,3]].
Then the output appends a column with its prediction for each pair,
e.g.                         [[1,2,1], [1,3,-1], [2,3,1]].
(1 means the first row is predicted bigger, -1 means the 2nd row.)}
}
\value{
A matrix with 3 columns: row1index, row2index, and the predicted greater index.
}
\description{
Generic function to use two sets of cues to predict which set has a higher criterion.
}

