% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictAlternative}
\alias{predictAlternative}
\title{Generic function to use two rows of cues to predict which row has a higher criterion.}
\usage{
predictAlternative(object, test_data, row_pairs = NULL)
}
\arguments{
\item{object}{The object that implements predictAlternative, e.g. a ttb model.}

\item{test_data}{The matrix of data to predict on.  As with predict, columns
must match those used for fitting.}

\item{row_pairs}{The optionalrow_pairs is a matrix of pairs of column indices
you want to have predicted.  By default (if no list is given), it will use
all unique pairs.
E.g. for 3 rows, it will use [[1,2], [1,3], [2,3]].}
}
\value{
A matrix with 3 columns: row1index, row2index, and the prob row1 greater.
 The first two columns are the row_pairs (provided as input or all).
 The 3rd column is the model's predicted probability (0 to 1) that the first
 row index has a larger criterion than the 2nd row index.  0.5 is a tie.

 For example:
                             [[1,2,1], [1,3,0], [2,3,1]].
 Means:
 Between row 1 and 2, there is probabily 1 that row 1 is bigger.
 Between row 1 and 3, there is probabily 0 that row 1 is bigger.
    (That is, it predicts that row 3 is bigger.)
Between row 2 and 3, there is probabily 1 that row 2 is bigger.
}
\description{
Generic function to use two rows of cues to predict which row has a higher criterion.
}

