% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregators.R
\name{predictAlternativeWithCorrect}
\alias{predictAlternativeWithCorrect}
\title{Generates a matrix of correct values an predictions among alternatives.}
\usage{
predictAlternativeWithCorrect(fitted_heuristic_list, test_data,
  rowPairs = NULL)
}
\arguments{
\item{fitted_heuristic_list}{List of heuristics that implement the generic function
predictAlternative, e.g. ttbBinModel.}

\item{test_data}{Data to try to predict; must match columns in fit.}

\item{rowPairs}{An optional matrix where the first two columns are the pairs
of row indices to use in the test_data.  If not set, all pairs will be used.}
}
\value{
Same matrix as predictAlternative but with columns on correctness
}
\description{
Generates a matrix of correct values an predictions among alternatives.
}
\seealso{
\code{\link{predictAlternative}}
}

