% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_pair_apply.R
\name{predictPair}
\alias{predictPair}
\title{Predict which of a pair of rows has a higher criterion.}
\usage{
predictPair(row1, row2, object)
}
\arguments{
\item{row1}{The first row of data.  The cues object$cols_to_fit will be
passed to the heuristic.}

\item{row2}{The second row of data.  The cues object$cols_to_fit will be
passed to the heuristic.}

\item{object}{The object that implements predictPairInternal, e.g. a fitted
ttbModel or logRegModel.}
}
\value{
A number in the set {-1, 0, 1}, where 1 means row1 is predicted to
  have a greater criterion, -1 means row2 is greater, and 0 is a tie.
}
\description{
Assumes the object implements predictProbInternal and has $cols_to_fit.
}

