% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{predictRoot}
\alias{predictRoot}
\title{Generic function to predict which of a pair of rows has a higher criterion.}
\usage{
predictRoot(object, row1, row2)
}
\arguments{
\item{object}{The object that implements predictPair, e.g. a ttb model.}

\item{row1}{The first row of cues (from cols_to_fit columns, based on object).}

\item{row2}{The second row}
}
\value{
A value from 0 to 1, representing the probability that row1's criterion
  is greater than row2's criterion.
}
\description{
Generic function to predict which of a pair of rows has a higher criterion.
}

