% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heuristic_classes.R
\name{ttbModel}
\alias{ttbModel}
\title{Take The Best}
\usage{
ttbModel(train_data, criterion_col, cols_to_fit, reverse_cues = TRUE)
}
\arguments{
\item{train_data}{Training/fitting data as a matrix or data.frame.}

\item{criterion_col}{The index of the colum in train_data that has the criterion.}

\item{cols_to_fit}{A vector of column indices in train_data, used to fit the criterion.}

\item{reverse_cues}{Optional parameter to reverse cues as needed.  By default,
the model will reverse the cue values for cues with cue validity < 0.5, so a cue
with validity 0 becomes a cue with validity 1.
Set this to FALSE if you do not want that, i.e. the cue stays validity 0.}
}
\value{
An object of \code{\link[base]{class}} ttbModel.  This is a list containing at least the following components:
  \itemize{
   \item "cue_validities": A list of cue validities for the cues in order of cols_to_fit.
  }
}
\description{
An implementation of the Take The Best heuristic.
It sorts cues in order of \code{\link{cueValidity}}, making a decision based on the first cue that
discriminates (has differing values on the two objects).
}
\details{
It does NOT implement predict on purpose.
Developer TODO: Have TTB reverse a cue with validity < 0.5.
}
\examples{
## Fit column (5,4) to column (1,0), having validity 1.0, and column (0,1), validity 0.
ttb <- ttbModel(matrix(c(5,4,1,0,0,1), 2, 3), 1, c(2,3))
predictAlternative(ttb, matrix(c(5,4,1,0,0,1), 2, 3))
## But this input results in an incorrect prediction:
predictAlternative(ttb, matrix(c(5,4,0,1,0,1), 2, 3))
}
\seealso{
\code{\link{predictAlternative.ttbModel}} (via \code{\link{predictAlternative}}) for prediction.

Wikipedia's entry on \url{http://en.wikipedia.org/wiki/Take-the-best_heuristic}.
}

